materials2/go_cgt_app/requirements.txt
Pillow
graphviz
